# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proxsis Workspace Web API Documentation",
        "description": "Every single endpoint are documented",
        "contact": {
            "email": "kokonur1020@gmai.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users/leave_approvals/{user_id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get Leave Approvals",
                "description": "Get Leave Approvals",
                "operationId": "leaveApprovalStatistics",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Leave Approvals Success"
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/attendance_approvals/{user_id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get Manual Attendance Approvals",
                "description": "Get Manual Attendance Approvals",
                "operationId": "attendanceApprovalStatistics",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Attendance Approvals Success"
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}